# Improve performance of sys.sp_column_privileges by removing an unutilized check