### Description

Improve performance of sys.sp_column_privileges by removing an unutilized check. 

### Issues Resolved

Task: BABEL-5353
Signed-off-by: Shalini Lohia <lshalini@amazon.com>

### Test Scenarios Covered ###

Performance Testing
20 login, 50 databases, each with 10 tables in dbo and 10 schema with 10 tables each making it 5.5k tables 
Each database has 20 users in each database making it 500+ users
Each database has 10 new roles.

```
1> select count(*) from sp_column_privileges_view
2> go

 4748 rows (1469 ms)


1> select * from sp_table_privileges_view
2> go

585 rows (1630 ms)
```


| query | Execution time before [with func](ms) | Execution time after [with func, current change](ms)| Execution time before [with join](ms)|Execution time before [without join, it returns incorrect value]|
|-------| -------| ------| ------| ------|
|EXEC sp_column_privileges @table_name = 'db1_table3' | 2903 | 1453| 1927| 1338|
|EXEC sp_column_privileges @table_name = 'db1_table3', @table_qualifier = 'db1', @COLUMN_NAME='column1' |  2900 | 1442| 1812| 1290|
|CREATE PROCEDURE usp_GetColumnPrivileges @TableName NVARCHAR(128) AS begin EXEC sys.sp_column_privileges @table_name = @TableName; end ; usp_GetColumnPrivileges @TableName = 'db1_table3'; | 2893 | 1454| 1860| 1339|
|EXEC sp_table_privileges @table_name = 'db1_table3' | 1590 | 1590| 1473 | 1435|


* **Use case based -**
* **Boundary conditions -**
* **Arbitrary inputs -**
* **Negative test cases -**
* **Minor version upgrade tests -**
* **Major version upgrade tests -**
* **Performance tests -**
* **Tooling impact -**
* **Client tests -**